Alteração Double para BigDecimal

+++ src/main/java/com/commerce/desafiodev/model/Numero.java
@@ -1,11 +1,13 @@
 package com.commerce.desafiodev.model;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 public class Numero {
-	private Double[] numbers;
+	private BigDecimal[] numbers;
 	
 }
