+++ src/main/java/com/commerce/desafiodev/repository/NumeroRepository.java
@@ -2,6 +2,7 @@ package com.commerce.desafiodev.repository;
 
 import com.commerce.desafiodev.model.ConjuntoNumero;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,5 +13,5 @@ import org.springframework.stereotype.Repository;
 public interface NumeroRepository extends JpaRepository<ConjuntoNumero, Integer> {
 	
 	@Query(value = "SELECT NUMERO FROM CONJUNTO_NUMERO", nativeQuery = true)
-	public List<Double> findNumber();
+	public List<BigDecimal> findNumber();
 }
